+++ src/constants/index.ts
@@ -1,3 +1,5 @@
 export * from './storageKeys';
 export * from './clinical';
 export * from './audioPhrases';
+export * from './goldenData';
+