+++ src/engine/kinematics/index.ts
@@ -4,3 +4,5 @@ export * from './parallaxCheck';
 export * from './poseDetector';
 export * from './occlusionHeuristics';
 export * from './rulesEngine';
+export * from './referenceDataLoader';
+